+++ besch/groundobj_besch.h
@@ -46,7 +46,7 @@ public:
 
 	bool is_allowed_climate( climate cl ) const { return ((1<<cl)&allowed_climates)!=0; }
 
-	const bild_besch_t *get_bild(int season, int phase) const  	{
+	const bild_besch_t *get_bild(uint8 season, uint16 phase) const  	{
 		return get_child<bildliste2d_besch_t>(2)->get_bild(phase, season);
 	}
