+++ besch/way_obj_besch.h
@@ -61,7 +61,7 @@ public:
 
 	image_id get_front_slope_image_id(hang_t::typ hang) const
 	{
-		int nr;
+		uint16 nr;
 		switch(hang) {
 			case 4:
 				nr = 0;
@@ -101,7 +101,7 @@ public:
 
 	image_id get_back_slope_image_id(hang_t::typ hang) const
 	{
-		int nr;
+		uint16 nr;
 		switch(hang) {
 			case 4:
 				nr = 0;
